inquiry

+++ inquiry_register/models/inquiry_register.py
@@ -6,7 +6,7 @@ class InquiryRegister(models.Model):
 
     name = fields.Char(string='Parent Name')
     student_name = fields.Char(string='Student Name')
-    child_age = fields.Integer(string='Child Age')
+    child_age = fields.Float(string='Child Age')
     phone_number = fields.Char(string='Phone Number')
     email = fields.Char(string='Email')
     purpose = fields.Text(string='Purpose')
